% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoEstCont.R
\name{autoEstCont}
\alias{autoEstCont}
\title{Automatically calculate the contamination fraction}
\usage{
autoEstCont(
  sc,
  nMarks = 50,
  tfidfMin = 0.8,
  soupMin = 1e-04,
  rhoMax = 1,
  FDR = 0.1,
  minCnts = 1,
  doPlot = TRUE,
  ...
)
}
\arguments{
\item{sc}{The SoupChannel object.}

\item{nMarks}{How many marker genes to use. These must also pass the tfidf cut-off and be highly expressed in the soup, so there may end up being fewer than this number.}

\item{tfidfMin}{Minimum value of tfidf to accept for a marker gene.}

\item{soupMin}{Minimum value of expression in the background to accept for marker.}

\item{rhoMax}{What contamination fraction is the maximum that is plausible.  Must be a value between 0 and 1.}

\item{FDR}{False discovery rate for statistical tests.}

\item{minCnts}{Don't use estimates with fewer than this many counts.}

\item{doPlot}{Create a plot summarising the estimation?}

\item{...}{Passed to density.}
}
\description{
The idea of this method is that genes that are highly expressed in the soup and are marker genes for some population can be used to estimate the background contamination.  Marker genes are identified using the tfidf method (see \code{\link{quickMarkers}}).  The contamination fraction is then calculated at the cluster level for each of these genes and clusters are then aggressively pruned to remove those that give implausible estimates.
}
\details{
The pruning of implausible clusters is based on several parameters.  Firstly, clusters are excluded for estimation on a gene-by-gene basis using \code{\link{estimateNonExpressingCells}} as would be done had the genes been specified by the user.  Any cluster whose estimation for rho is not significantly less than \code{rhoMax} (as measured by a Poisson test) is also thrown out (with false discovery rate \code{FDR}).

Finally, clusters are excluded if they number of observed counts is less than \code{minCnts}.  The default (where \code{minCnts=1}) essentially prevents an excess of estimates of zero contamination skewing the estimation.

An estimate of the global contamination rate is then provided based on the value of rho with the highest density in the estimates using the pruned cluster/gene pairs.  The logic behind this is that  clusters that truly estimate the contamination fraction will cluster around the true value, while erroneous estimates will be spread out across the range (0,1) without a 'preferred value'.

Note that the interpretation of the tf-idf cut-off in this context is that a cut-off t implies that a marker gene has the property that geneFreqGlobal < exp(-t/geneFreqInClust).
}
