% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseGeneSets.R
\name{chooseGeneSets}
\alias{chooseGeneSets}
\title{Selects sets of genes from a preselected \code{cadidateList} to estimate the contamination fraction of a soupChannel object.}
\usage{
chooseGeneSets(
  sc,
  inferredGenes = NULL,
  candidateList,
  top_n = 20,
  prefixDrop,
  ...
)
}
\arguments{
\item{sc}{A SoupChannel object.}

\item{inferredGenes}{A gene table passed from \code{\link{inferNonExpressedGenes}}. If \code{NULL}, the function is run on the SoupChannel object.}

\item{candidateList}{A list of pre-selected candidate gene sets to cross-reference with the inferred non-expressed genes.}

\item{top_n}{How many genes to consider from the top of the inferredGenes table.}

\item{prefixDrop}{A character scalar indicating the prefix to drop from the gene names.}

\item{...}{Passed to \code{\link{inferNonExpressedGenes}}.}
}
\value{
A dataframe with each of the tested gene sets and specific genes. They are listed in order by the estimated global contamination fraction and the summed extremity score for all the genes in the set.
}
\description{
The purpose of this function is to identify gene sets and specific genes in the candidate list that are also good candidates for estimating soup contamination.
}
\details{
The curated sets can be compared and passed to \code{\link{calculateContaminatingFraction}} to estimate the Soup contamination in each channel.
}
