% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanPlotFunctions.R
\name{cleanMarkerMap}
\alias{cleanMarkerMap}
\title{Cleans the ggplot object generated by \code{plotMarkerMap} for presentation.}
\usage{
cleanMarkerMap(gg_plot)
}
\arguments{
\item{gg_plot}{ggplot2 object passed from \code{plotMarkerDistribution}}
}
\value{
A ggplot2 containing the plot.
}
\description{
Given some reduced dimensional representation of the data (such as UMAP or tSNE) that has been calculated however you would like, this provides a way to visualise how likely a set of genes are to be soup derived on that map.  That is, given a set of genes, this function calculates how many counts would be expected if that droplet were nothing but soup and compares that to the observed count.  This is done via a log2 ratio of the two values.  A Poisson test is performed and points that have a statistically significant enrichment over the background (at 5% FDR) are marked.
}
