% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCleanedSeurat.R
\name{createCleanedSeurat}
\alias{createCleanedSeurat}
\title{Convert multiple strained channels into a Seurat object}
\usage{
createCleanedSeurat(cleanedChannels, scale.factor = 10000,
  channelCellPrefix = "", ...)
}
\arguments{
\item{cleanedChannels}{A list of channels that have been run through \code{\link{strainChannel}}.}

\item{...}{Extra parameters passed to \code{CreateSeuratObject}.}

\item{channelPrefixes}{A vector of the same length as \code{cleanedChannels} giving a prefix to add to each cell before merging.}
}
\value{
A normalised Seurat object.
}
\description{
Constructs a Seurat object from the channels given and normalises the result.
}
