% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estGrouped.R
\name{estGrouped}
\alias{estGrouped}
\title{Estimates contamination from a group of cells}
\usage{
estGrouped(obsCnts, expCnts, nUMIs, grpFac, logRho = FALSE)
}
\arguments{
\item{obsCnts}{The observed number of counts that can confidently be assigned to background for each cell.}

\item{expCnts}{The expected number of counts for each cell if the cell were to contain nothing but background.}

\item{nUMIs}{The total UMIs for each cell.}

\item{grpFac}{A vector defining how cells are to be grouped together for calculating rho.}

\item{logRho}{Should a log10 transformation be applied to the estimate of rho?}
}
\value{
A data.frame with estimates of rho by group, along with confidence bounds and some meta-data.
}
\description{
Intended for internal use only.  Given expected and observed soup counts for a group of cells and a vector defining how to group, returns the maximum likelihood estimator of the contamination fraction for each group.
}
