% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCells.R
\name{findCells}
\alias{findCells}
\title{Estimates which droplets contain a cell}
\usage{
findCells(sc, nonExpressedGeneList, useToEst = NULL, minUMIs = 100,
  minGenes = 50, contCut = 0.3)
}
\arguments{
\item{sc}{A SoupChannel object for which the table of droplets has been retained.  That is, \code{\link{estimateSoup}} was run with \code{keepDroplets=TRUE}.}

\item{useToEst}{Which cells to use which sets of genes to estimate contamination (see \code{\link{calculateContaminationFraction}})}

\item{minUMIs}{Don't test any droplet with fewer than this many UMIs.}

\item{minGenes}{Don't test any droplet with fewer than this many detected genes.}

\item{contCut}{The maximum allowable contamination fraction for a droplet containing a cell.}

\item{soupProfile}{The expression profile of the soup for this channel, as calculated by \code{\link{estimateSoup}}.}

\item{nonExpressedGenes}{Genes to use for estimating contamination fraction (see \code{\link{calculateContaminationFraction}})}
}
\value{
A data.frame with contamination fraction estimates for each droplet and a p-value and FDR for rejecting the NULL hypothesis and accepting this droplet as a cell.
}
\description{
This function attempts to identify which droplets contain cells based on the level of contamination in each cell.  For this to work well, there must be sufficient power to estimate the contamination fraction in each droplet, which is usually not the case unless spike-ins have been used.  To determine which droplets contain cells, a binomial test (with FDR correction) is performed against the NULL hypothesis that the contamination rate is equal to or greater than \code{contCut}.
}
