% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifyExpressingCells.R
\name{identifyExpressingCells}
\alias{identifyExpressingCells}
\title{Identify expressing cells}
\usage{
identifyExpressingCells(sc, channelName, nonExpressedGeneList,
  excludeMethod = c("pCut", "thresh"), exCut = 0.05)
}
\arguments{
\item{sc}{A SoupChannel or SoupChannelList object.  If a SoupChannelList object, channelName must be given and must be the name of a channel.}

\item{channelName}{The name of a channel to use if \code{sc} is a SoupChannelList object.}

\item{nonExpressedGeneList}{A list containing sets of genes which can be assumed to be non-expressed in a subset of cells (see details).}

\item{excludeMethod}{Which method to use to exclude cells (see details).}

\item{exCut}{Cut-off used by \code{excludeMethod} (see details).}
}
\value{
A boolean matrix with cells that definitively express the set of genes given by the row name set to FALSE.  Usually passed to \code{\link{calculateContaminationFraction}} via the \code{useToEst} argument.
}
\description{
This function identifies which cells truly express different sets of genes and which only appear to express it due to ambient mRNA contamination.  A list containing genes that can be expected to be non-expressed in at least one cell must be given.  That is, this list contains sets of genes (e.g. Haemoglobin genes) that can be assumed to be completely absent from certain cells.  For example, haemoglobin genes should not be expressed in anything that is not a red blood cell.
}
\details{
The different methods to determine which cells to exclude which genes for are:
  pCut - Exclude any cell where we can reject the null hypothesis (Poisson test) that the contamination fraction is less than or equal to 1 for a set of genes at the \code{exCut} level.
  thresh - Exclude any cell for which the estimate for contamination fraction exceeds \code{exCut}.
}
\seealso{
\code{\link{calculateContaminationFraction}}
}
