% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{initProgBar}
\alias{initProgBar}
\title{Create Seurat style progress bar}
\usage{
initProgBar(min, max, ...)
}
\arguments{
\item{min}{Minimum value of parameter.}

\item{max}{Maximum value of parameter.}

\item{...}{Passed to \code{\link{txtProgressBar}}}
}
\value{
A txtProgressBar object to use updating progress.
}
\description{
Creates progress bar that won't ruin log files and shows progress towards 100%.
}
