% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolateCellContamination.R
\name{interpolateCellContamination}
\alias{interpolateCellContamination}
\title{Get cell specific rho from binned values}
\usage{
interpolateCellContamination(sc, channelName, useGlobal = FALSE)
}
\arguments{
\item{sc}{A \code{SoupChannel} or \code{SoupChannelList} object.}

\item{channelName}{The name of the channel to use if \code{sc} is a \code{SoupChannelList} object.}

\item{useGlobal}{If TRUE, set all cells to the global estimate of the contamination fraction.}
}
\value{
A modified version of \code{sc} with cell estimates of contamination stored in \code{rhos}.  If \code{sc} is a \code{SoupChannelList}, the named channel is updated in this way.
}
\description{
Interpolates rho at each cell based on each cell's total number of UMIs from estimates in bins.
}
