% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PBMC}
\alias{PBMC}
\title{PBMC data}
\source{
\url{https://support.10xgenomics.com/single-cell-gene-expression/datasets/2.1.0/pbmc4k}
}
\usage{
data(PBMC)
}
\description{
Collection of bits of data relating to the 10X PBMC 4K data.
}
\details{
This collection includes \code{PBMC_DR}, \code{PBMC_cellBarcodes}, and \code{PBMC_tod}.  See individual help for more details.
}
