% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PBMC_cellBarcodes}
\alias{PBMC_cellBarcodes}
\title{Cell barcodes for PBMC data}
\format{Vector of cellular barcodes.}
\source{
\url{https://support.10xgenomics.com/single-cell-gene-expression/datasets/2.1.0/pbmc4k}
}
\usage{
data(PBMC)
}
\description{
Which barcodes contain cells (as determined by cellranger) in the PBMC data.  Created by running:
\code{PBMC_cellBarcodes = colnames(Read10X('filtered_gene_bc_matrices/GRCh38/'))}
}
