% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PBMC_SCL}
\alias{PBMC_SCL}
\title{Basic Soup Channel List for PBMC data}
\format{One SoupChannelList object}
\usage{
data(PBMC_SCL)
}
\description{
Basic output of \code{\link{load10X}} on PBMC data
}
