% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PBMC_tod}
\alias{PBMC_tod}
\title{Raw PBMC data}
\format{Sparse matrix with rows representing genes, columns cells and entries number of unique molecular identifiers detected.}
\source{
\url{https://support.10xgenomics.com/single-cell-gene-expression/datasets/2.1.0/pbmc4k}
}
\usage{
data(PBMC)
}
\description{
Raw PBMC data used in the vignette.  This is the full matrix including all droplets in sparse format.  Created from the PBMC 4k data using:
\code{PBMC_tod = Read10X('raw_gene_bc_matrices/GRCh38/')}
}
