% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotCandidateMarkerGenes}
\alias{plotCandidateMarkerGenes}
\title{Plot non-expressed genes for channel}
\usage{
plotCandidateMarkerGenes(sc, channel, n = 20, ...)
}
\arguments{
\item{sc}{A SoupChannel or SoupChannelList object.}

\item{n}{Number of candidate genes to plot.}

\item{...}{Parameters passed to \code{\link{plotMarkerDistribution}}.}

\item{channelName}{The name of the channel to plot.  Ignored if \code{sc} is of class \code{SoupChannel}.}
}
\value{
A ggplot2 object containing the plot.
}
\description{
A helper function to plot the top N non-expressed genes for a channel.
}
