% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotChangeMap}
\alias{plotChangeMap}
\title{Plot maps comparing corrected/raw expression}
\usage{
plotChangeMap(scl, geneSet, DR, dataType = c("binary", "counts",
  "expression", "ratio"), logData = TRUE,
  includePanels = c("Uncorrected", "CorrectedExpression",
  "CorrectedCounts"))
}
\arguments{
\item{scl}{SoupChannelList object.}

\item{geneSet}{A vector with the names of the genes to aggregate and plot evidence for.}

\item{DR}{A data.frame, with rows named by unique cell IDs (i.e., <ChannelName>___<Barcode>) the first two columns of which give the coordinates of each cell in some reduced dimension representation of the data.}

\item{dataType}{How should data be represented.  Binary sets each cell to expressed or not, counts converts everything to counts, expression converts everything to expression.  Ratio plots the ratio of everything to uncorrected (and uncorrected is shown as z-scaled expression).}

\item{logData}{Should we log data?}

\item{includePanels}{Which of 'Uncorrected','CorrectedExpression' and 'CorrectedCounts' to include.  Also sets plot ordering.}
}
\value{
A ggplot2 containing the plot.
}
\description{
Given some reduced dimensional representation of the data (such as UMAP or tSNE) that has been calculated however you would like, this provides a way to visualise how the expression of a geneSet changes after soup correction.
}
