% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotChannelContamination}
\alias{plotChannelContamination}
\title{Plot contamination fraction for channel}
\usage{
plotChannelContamination(rhos, showErrorBars = TRUE)
}
\arguments{
\item{showErrorBars}{Should error bars showing the 95 percent confidence interval on the estimate of rho be included?}

\item{rho}{Output from \code{\link{calculateContaminationFraction}}.}
}
\value{
A ggplot2 object containing the plot.
}
\description{
Plots the contamination fraction as a function of nUMIs in a droplet.
}
