% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotChannelSummary}
\alias{plotChannelSummary}
\title{Plots a summary of the contamination for a channel}
\usage{
plotChannelSummary(cleanedChannel)
}
\arguments{
\item{cleanedChannel}{Output of \code{\link{strainChannel}} or \code{\link{strain10X}}.}
}
\value{
A ggplot2 object containing the plot.
}
\description{
Takes the output of \code{\link{strainChannel}} (or \code{\link{strain10X}}) and plots a summary.  This just pieces together the plots produced by \code{\link{plotChannelSummary}}, \code{\link{plotSoupCorrelation}}, \code{\link{plotMarkerDistribution}} and \code{\link{inferNonExpressedGenes}}.
}
