% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotDropletResolution}
\alias{plotDropletResolution}
\title{Plots UMI counts and Droplet status (used for correction, for soup or for none)}
\usage{
plotDropletResolution(sc, subsample = 1e+06)
}
\arguments{
\item{sc}{SoupChannel object.}

\item{subsample}{subsample to at most this number of droplets, use 0 for including all droplets}
}
\value{
A ggplot2 containing the plot.
}
\description{
Plots UMI counts and Droplet status (used for correction, for soup or for none)
}
