% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classFunctions.R
\name{print.SoupChannel}
\alias{print.SoupChannel}
\title{Print method for SoupChannel}
\usage{
\method{print}{SoupChannel}(x, ...)
}
\arguments{
\item{x}{A SoupChannel object.}

\item{...}{Currently unused.}
}
\value{
Nothing.  Prints message to console.
}
\description{
Prints a summary of a SoupChannel object.
}
