% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classFunctions.R
\name{print.SoupChannelList}
\alias{print.SoupChannelList}
\title{Print method for SoupChannelList}
\usage{
\method{print}{SoupChannelList}(scl)
}
\arguments{
\item{scl}{A SoupChannelList object.}
}
\description{
Prints a summary of a SoupChannelList object.
}
