% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMtx10X.R
\name{readMtx10X}
\alias{readMtx10X}
\title{Load the 10X data matrix}
\usage{
readMtx10X(dataDirs, chNames = names(dataDirs), isV3 = FALSE)
}
\arguments{
\item{chNames}{Names of each experiment to prepend to barcodes to make them unique.}

\item{isV3}{Are these cellranger v3.0.0 or greater outputs?  Can be a vector.}

\item{Vector}{of directories containing matrix.mtx(.gz) files to load.}
}
\value{
Matrix containing all data.
}
\description{
Loads 10X cellranger output from multiple directories and merge.
}
