% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeNoncorrectedDroplets.R
\name{removeNoncorrectedDroplets}
\alias{removeNoncorrectedDroplets}
\title{Removes droplets from the SoupChannel object not actually corrected}
\usage{
removeNoncorrectedDroplets(sc)
}
\arguments{
\item{sc}{A SoupChannel object}
}
\value{
A SoupChannel object.
}
\description{
Removes droplets from the SoupChannel object not actually corrected
}
\note{
This is a cleanup function
}
