% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classFunctions.R
\name{SoupChannel}
\alias{SoupChannel}
\title{Construct a SoupChannel object}
\usage{
SoupChannel(tod, toc, metaData = NULL, soupEstParams = list(), ...)
}
\arguments{
\item{tod}{Table of droplets.  A matrix with columns being each droplet and rows each gene.}

\item{toc}{Table of counts.  Just those columns of \code{tod} that contain cells.}

\item{metaData}{Meta data pertaining to the cells.  Optional.  Must be a data-frame with rownames equal to column names of \code{toc}.}

\item{soupEstParams}{Pramaters passed to \code{\link{estimateSoup}}.}

\item{...}{Any other named parameters to store.}
}
\value{
A SoupChannel object.
}
\description{
Creates a SoupChannel object that contains everything related to the soup estimation of a single channel.
}
\seealso{
SoupChannelList estimateSoup
}
