% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classFunctions.R
\name{SoupChannelList}
\alias{SoupChannelList}
\title{Construct a SoupChannelList object}
\usage{
SoupChannelList(channels, ...)
}
\arguments{
\item{channels}{A list of \code{\link{SoupChannel}} objects.  Must be uniquely named.}

\item{...}{Any other named parameter to store.}
}
\value{
A SoupChannelList object.
}
\description{
Creates a SoupChannelList object that contains everything related to the soup estimation for a collection of channels, usually one experiment.
}
\seealso{
SoupChannel estimateSoup
}
