% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikeInTheSoupChannel.R
\name{spikeInTheSoupChannel}
\alias{spikeInTheSoupChannel}
\title{Sample spike-in/soup Droplets from the Soup channel}
\usage{
spikeInTheSoupChannel(sc, target.spikein.droplet.count = 5000, seed = 1)
}
\arguments{
\item{sc}{A \code{SoupChannel} object.}

\item{target.spikein.droplet.count}{per-Droplet read depth to sample}

\item{seed}{Random seed, to ensure reproducibility.}
}
\value{
Spike-in / soup sampled droplets
}
\description{
Sample spike-in/soup Droplets from the Soup channel
}
\examples{
sc = load10X(system.file(package = "SoupX", "extdata/toyData"), keepDroplets=TRUE)
expected.count = round(median(Matrix::colSums(sc$toc)))
si = spikeInTheSoupChannel(sc,target.spikein.droplet.count= expected.count)
plot(sort(Matrix::colSums(si)))
removeNoncorrectedDroplets(sc)
}
