% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strain10X.R
\name{strain10X}
\alias{strain10X}
\title{Estimates background contamination and removes it from a 10X channel}
\usage{
strain10X(dataDirs, nonExpressedGeneList, soupRange = c(0, 10), ...)
}
\arguments{
\item{dataDirs}{Path to the 10X Cell Ranger output.  Should be the top level folder containing the html report.  Can be a vector, in which case each is processed in turn with the all other parameters unchanged.}

\item{nonExpressedGeneList}{Which genes to use to estimate soup (see \code{\link{calculateContaminationFraction}})}

\item{soupRange}{Which droplets to estimate the soup from (see \code{\link{estimateSoup}})}

\item{...}{Extra arguments passed to \code{\link{calculateContaminationFraction}}.}
}
\value{
The same as \code{\link{strainChannel}}
}
\description{
Takes the definition of cells from the "filtered counts" folder, then runs the standard workflow to estimate and remove contamination.
}
