% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strainCells.R
\name{strainCells}
\alias{strainCells}
\title{Remove background contamination from cell expression profile}
\usage{
strainCells(scl, groupVec = NULL, retainSparsity = FALSE)
}
\arguments{
\item{scl}{A SoupChannel or SoupChannelList object.}

\item{groupVec}{Vector named by unique droplet IDs of the format channelName___dropletBarcode that are unique across channels, with the entries indicating the name of the group each cell belongs to.  Cells in the same group are given an aggregated expression profile after background correction.  If NULL, all cells are corrected individually.  Any unmentioned cells are also corrected individually.}

\item{retainSparsity}{The MLE optimisation is much faster for non-sparse matrices when \code{groupVec} is not NULL.  By default the code will coerce a sparse matrix to be non-sparse in this case unless this flag is set to TRUE.}
}
\value{
A modified version of \code{scl} that has an extra matrix, scl$strainedExp, that contains the corrected expression matrix.
}
\description{
Given a table of counts composed of cells from multiple channels, estimates of the background contamination in each cell and the soup expression profile for each channel, calculates the true cell specific expression profile for each cell or a set of cells.  If background correction is being performed on cells all from the same channel, \code{soupProfiles} can be the output of \code{estimateSoup} for a single channel.
}
\seealso{
\code{\link{adjustCounts}}
}
