% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strainChannel.R
\name{strainChannel}
\alias{strainChannel}
\title{Estimates background contamination from a single sample}
\usage{
strainChannel(tod, cellIdxs, nonExpressedGeneList, soupRange = c(0, 10), ...)
}
\arguments{
\item{tod}{Table of counts (UMIs) where columns are samples and rows are genes.  Must include all droplets, not just those containing cells (e.g., raw_gene_bc_matricies for 10X data)}

\item{cellIdxs}{Indices for the columns of \code{toc} that contain cells.}

\item{nonExpressedGeneList}{Which genes to use to estimate soup (see \code{\link{calculateContaminationFraction}})}

\item{soupRange}{Which droplets to estimate the soup from (see \code{\link{estimateSoup}})}

\item{...}{Extra arguments passed to \code{\link{calculateContaminationFraction}}.}
}
\description{
Estimates the background contamination profile, cell-specific contamination fraction and the expression profile for each cell without contamination.
}
