% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustCounts.R
\name{adjustCounts}
\alias{adjustCounts}
\title{Adjust counts to remove soup}
\usage{
adjustCounts(scl, pCut = 0.01, verbose = TRUE)
}
\arguments{
\item{scl}{A SoupChannelList object for which rho has been calculated.}

\item{pCut}{The threshold for excluding counts.}

\item{verbose}{By default, this method prints out the 100 most removed genes by channel.  If this is set to FALSE, these are not printed.}
}
\value{
A modified version of \code{scl} with the corrected count matrix stored in atoc.
}
\description{
Explicitly drop counts from the count matrix in order of most likely to be soup to least likely until we have removed as many counts as we expect there to be soup molecules in each cell.  The actual removal criteria is to calculate the probability of each count being from the soup in each cell, pMol, and the probability of removing another soup molecule from this cell, pSoup, and keep sequentially removing counts until pSoup*pMol<pCut.  That is, pSoup makes sure we don't throw out many more than we expect to be present in each cell and pMol makes sure we don't throw out any molecule that is unlikely to be soup.
}
\seealso{
\code{\link{strainCells}}
}
