% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateContaminationFraction.R
\name{calculateContaminationFraction}
\alias{calculateContaminationFraction}
\title{Calculate the contamination fraction}
\usage{
calculateContaminationFraction(sc, channelName, nonExpressedGeneList,
  useToEst = NULL, cellGroups = NULL, tgtSoupCntsPerGroup = 1000,
  excludeMethod = c("pCut", "thresh"), exCut = 0.05)
}
\arguments{
\item{sc}{A SoupChannel or SoupChannelList object.  If a SoupChannelList object, channelName must be given and must be the name of a channel.}

\item{channelName}{The name of a channel to use if \code{sc} is a SoupChannelList object.}

\item{nonExpressedGeneList}{A list containing sets of genes which can be assumed to be non-expressed in a subset of cells (see details).}

\item{useToEst}{A boolean matrix of dimensions length(nonExpressedGeneList) x ncol(toc) indicating which gene-sets should not be assumed to be non-expressed in each cell.  Row names must correspond to the names of \code{nonExpressedGeneList}.  If NULL, this is estimated from the data (see details).}

\item{cellGroups}{A vector indicating which cells to group together when estimating rho (see details).}

\item{tgtSoupCntsPerGroup}{When automatically constructing groups, ensure that each group has roughly this many expected soup counts.}

\item{excludeMethod}{Which method to use to exclude cells (see details).}

\item{exCut}{Cut-off used by \code{excludeMethod} (see details).}
}
\value{
A modified version of \code{sc} with the entry \code{rhoGrouped} containing estimates of the contamination fraction for different groupings.  If \code{sc} is a SoupChannelList object the returned object has the channel named \code{channelName} modified to include the estimates of rho.
}
\description{
This function calculates the contamination fraction using the markers provided in groups of cells.  A list containing genes that can be expected to be non-expressed in at least one cell must be given.  That is, this list contains sets of genes (e.g. Haemoglobin genes) that can be assumed to be completely absent from certain cells.  For example, haemoglobin genes should not be expressed in anything that is not a red blood cell.  The contamination fraction is then calculated using these genes in groups of cells.
}
\details{
As well as sets of non expressed genes, the user must indicate which cells these genes are non-expressed in.  Continuing with the haemoglobin gene example, this would mean the user must indicate which cells contain red blood cells.  This information is given by the parameter \code{useToEst} which should be a boolean matrix with columns representing cells in toc and rows representing groups of genes defined in \code{nonExpressedGeneList}.  If this is set to NULL, the code will attempt to estimate which cells truly express each set of genes in \code{nonExpressedGeneList} based on the method given in \code{cellExpressionMethod}.

Unless a large list of non expressed genes can be given, there is usually insufficient power to estimate the contamination fraction separately for each cell.  Therefore, the user should supply a vector to \code{cellGroups} that can be used to group together cells when estimating the contamination fraction.  If set to NULL, the code will group together cells with a similar number of total UMIs in groups such that the expected number of soup counts in each group is roughly equal to \code{tgtSoupCntsPerGroup}.

The different methods to determine which cells to exclude which genes for are:
  pCut - Exclude any cell where we can reject the null hypothesis (Poisson test) that rho<=1 for a set of genes at the \code{exCut} level.
  thresh - Exclude any cell for which the estimate for rho exceeds \code{exCut}.
}
