% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCleanedSeurat.R
\name{createCleanedSeurat}
\alias{createCleanedSeurat}
\title{Convert multiple strained channels into a Seurat object}
\usage{
createCleanedSeurat(scl, scale.factor = 10000, ...)
}
\arguments{
\item{scl}{A \code{SoupChannelList} object on which \code{\link{strainCells}} has been run.}

\item{scale.factor}{The scale factor used in normalising data to log(1+scale.factor*x)}

\item{...}{Extra parameters passed to \code{CreateSeuratObject}.}
}
\value{
A normalised Seurat object.
}
\description{
Constructs a Seurat object from the channels given and normalises the result.
}
