% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateSoup.R
\name{estimateSoup}
\alias{estimateSoup}
\title{Get expression profile of soup}
\usage{
estimateSoup(tod, soupRange = c(0, 10))
}
\arguments{
\item{tod}{Table of UMIs for all droplets.}

\item{soupRange}{Droplets with total UMI count in this range (excluding endpoints) are used to estimate soup.}
}
\value{
A data.frame with the soup profile and confidence limits for all genes.
}
\description{
Uses the empty droplets in the range provided to calculate the expression profile of the soup under the assumption that these droplets only contain background.
}
