% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PBMC_cellBarcodes}
\alias{PBMC_cellBarcodes}
\title{PBMC 4K cellular barcodes}
\format{\code{PBMC_cellBarcodes} is a vector containing the barcodes for droplets containing cells.}
\source{
\url{https://support.10xgenomics.com/single-cell-gene-expression/datasets/2.1.0/pbmc4k}
}
\usage{
data(PBMC_cellBarcodes)
}
\description{
A vector of cellular barcodes
}
\details{
This vector was created by running \code{PBMC_cellBarcodes = colnames(Seurat::Read10X('filtered_gene_bc_matrices/GRCh38/'))}.
}
\keyword{datasets}
