% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotChannelContamination}
\alias{plotChannelContamination}
\title{Plot contamination fraction for channel}
\usage{
plotChannelContamination(sc, channelName, showErrorBars = TRUE)
}
\arguments{
\item{sc}{A \code{SoupChannel} object on which \code{\link{calculateContaminationFraction}} has been run.  Alternatively, a \code{SoupChannelList} object containing such a channel.}

\item{channelName}{The name of the channel to use if \code{sc} is a \code{SoupChannelList}}

\item{showErrorBars}{Should error bars showing the 95 percent confidence interval on the estimate of rho be included?}
}
\value{
A ggplot2 object containing the plot.
}
\description{
Plots the contamination fraction as a function of nUMIs in a droplet.
}
