% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFunctions.R
\name{plotSoupCorrelation}
\alias{plotSoupCorrelation}
\title{Plot correlation of expression profiles of soup and aggregated cells}
\usage{
plotSoupCorrelation(toc, soupProfile)
}
\arguments{
\item{toc}{Table of UMIs for all cells.}

\item{soupProfile}{The expression profile of the soup (output from \code{\link{estimateSoup}})}
}
\value{
A ggplot2 object containing the plot.
}
\description{
Calculates an expression profile by aggregating counts across all cells and plots this (on a log10 scale) against the expression profile of the soup.
}
