% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classFunctions.R
\name{print.SoupChannel}
\alias{print.SoupChannel}
\title{Print method for SoupChannel}
\usage{
\method{print}{SoupChannel}(sc)
}
\arguments{
\item{sc}{A SoupChannel object.}
}
\description{
Prints a summary of a SoupChannel object.
}
