% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classFunctions.R
\name{print.SoupChannelList}
\alias{print.SoupChannelList}
\title{Print method for SoupChannelList}
\usage{
\method{print}{SoupChannelList}(x, ...)
}
\arguments{
\item{x}{A SoupChannelList object.}

\item{...}{Currently unused.}
}
\description{
Prints a summary of a SoupChannelList object.
}
